Add bar labels

+++ data_skew/core.py
@@ -41,9 +41,9 @@ def get_filename(input_file):
 # allow user to specify input file name as command line argument
 parser = OptionParser()
 parser.add_option("-i", "--input-file", dest="input_file", action="store", type="string", help="The input filepath.")
-parser.add_option("-a", "--age", dest="age", action="store", type="int", help="The age to examine. Default is the max age in the input file.")
+parser.add_option("-a", "--age", dest="age", action="store", type="float", help="The age to examine. Default is the max age in the input file.")
 parser.add_option("-z", "--z-outlier-threshold", dest="z_outlier_threshold", action="store", type="int", default=3.5, help="A value whose modified z-score has an absolute value above this threshold is considered an outlier. Default is 3.5.")
-parser.add_option("-e", "--extreme-outlier-threshold", dest="extreme_outlier_threshold", action="store", type="int", default=7, help="A value whose modified z-score has an absolute value above this threshold is considered an extreme outlier. Default is 7.")
+parser.add_option("-e", "--extreme-outlier-threshold", dest="extreme_outlier_threshold", action="store", type="float", default=7, help="A value whose modified z-score has an absolute value above this threshold is considered an extreme outlier. Default is 7.")
 
 
 (options, args) = parser.parse_args()
